FIx targetting system and implemented enemy & soldier UI

+++ Assets/Scripts/ImmuneCell.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Immune Cell", menuName = "Immune Cell")]
 public class ImmuneCell : ScriptableObject
 {
-    public double health;
-    public double atk;
+    public int health;
+    public int atk;
     public double atkRadius;
     public float movSpd;
 }
